requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.3
numpy==2.1.2
requests==2.32.3
